move statedata code into qt component
